good_instruction fix for oracle + data type fix

+++ uarchsim/writeback.cc
@@ -161,7 +161,7 @@ void pipeline_t::writeback(unsigned int lane_number) {
       REN->set_complete(PAY.buf[index].AL_index);
       if (PAY.buf[index].vp_eligible==1 && !REN->is_vp_perfect()){
       
-	      REN->set_vpq_value(PAY.buf[index].vpq_index,PAY.buf[index].C_value.dw);
+	      REN->set_vpq_value(PAY.buf[index].vpq_index,(int64_t)PAY.buf[index].C_value.dw);
       
       }
       // FIX_ME #16 END
